+++ launcher/source/menu.cpp
@@ -96,7 +96,7 @@ static void* UpdateGUI(void* arg)
 			if (exitRequested) {
 				for(i = 0; i < 255; i += 15) {
 					Window->Draw();
-					Menu_DrawRectangle(0, 0, screenwidth, screenheight, (GXColor){0, 0, 0, i}, 1);
+					Menu_DrawRectangle(0, 0, screenwidth, screenheight, (GXColor){0, 0, 0, (u8)i}, 1);
 					Menu_Render();
 				}
 				ExitApp();
